### 2024-07-04 7bc6eb7
change Time for destriping
