+++ paper/paper.bib
@@ -8,6 +8,7 @@
   publisher    = {Nature Methods},
   volume       = 9,
   pages        = "676-682",
+  doi          = {10.1038/nmeth.2019},
   url          = {https://doi.org/10.1038/nmeth.2019}
 }
 
@@ -21,7 +22,8 @@
   publisher    = {IEEE Transactions on Image Processing},
   volume       = 7,
   number       = 1,
-  pages        = "27-41"
+  pages        = "27-41",
+  doi          = {10.1109/83.650848}
 }
 
 @software{pyVSNR,
@@ -46,5 +48,6 @@
   publisher    = {IEEE Image Processing},
   volume       = 21,
   number       = 10,
-  pages        = "4420-4430"
+  pages        = "4420-4430",
+  doi          = {10.1109/TIP.2012.2206037}
 }
