### 2024-07-04 9967007
add Usage and Contributing sections
