### 2024-07-04 2e0f91a
change Time for destriping
