+++ doc/intro.rst
@@ -62,17 +62,15 @@ Here is an example of processing times associated with the example provided `her
    * - bkg_removal
      - 40
    * - destriping
-     - 480*
-   * - registration**
+     - 700*
+   * - registration
      - 24
    * - intensity_rescaling
      - 24
    * - resampling
      - 11
 
-(*) **destriping** calculation performed on a GPU Nvidia A-100, with images processed one by one.
-
-(**) **registration_calculation** and **registration_transformation** cumulative time
+(*) 120s with a GPU Nvidia A-100.
 
 Install
 -------
