### 2024-07-04 a1b0193
remove limitation nproc=1 for destriping
