### 2024-07-04 #6954 @patquem
Hi @Kevin-Mattheus-Moerman ,

Thank you very much for the registration in the JOSS reviewing process and for your comments.

* The DOIs have been updated.

* The contributing section is now included in the README and links to the documentation (hoping this is okay with you).

Regarding the "Five most similar historical JOSS papers," they seem quite different from pyStack3D, but I will let the editor form their own opinion. I understand that they are quite busy at the moment and I hope that one will soon be available to start the reviewing process.

Regarding the reviewers, among the list of people likely to review for JOSS in connection with this paper, I had noted: Sklumpe , pr4deepr and two other people who are not currently JOSS reviewers but work on processes similar to those proposed in pystack3d: gleb-shtengel, kleinerELM.
