### 2024-08-01 #6954 @patquem
@mstimberg

>  If you have any suggestions for potential reviewers ...

I made some suggestions at the the end of my message https://github.com/openjournals/joss-reviews/issues/6954#issuecomment-2208724619
