### 2024-08-01 #6954 @mstimberg
> I made some suggestions at the the end of my message [#6954 (comment)](https://github.com/openjournals/joss-reviews/issues/6954#issuecomment-2208724619)

Ah, I overlooked this, thanks!
