### 2024-08-06 #7079 @editorialbot
[REVIEW]: PyStack3D: A python package for fast image stack correction
**Submitting author:** @patquem (<a href="http://orcid.org/0009-0008-6936-1249">Patrick Quéméré</a>)
**Repository:** https://github.com/CEA-MetroCarac/pystack3d
**Branch with paper.md** (empty if default branch):
**Version:** v2024.1
