### 2024-08-06 #7079 @mstimberg
👋🏼 @patquem, @kasasxav, @xiuliren, @sklumpe this is the review thread for the paper. All of our communications will happen here from now on.
at the top of a new comment in this thread.

Please feel free to ping me (@mstimberg) if you have any questions/concerns. Thanks again for helping out!
