### 2024-08-08 #7079 @xiuliren
- I have tried it with my own FIB-SEM data. The usage requires that the image sequence be named as a specific pattern and is not super user-friendly. I can write a simple script to transform it, but it would be much better to have an interface for loading a normal image stack, such as tif or HDF5 format.
- We have also published a paper for FIB-SEM image preprocessing. Please cite it! [High-fidelity Image Restoration of Large 3D Electron Microscopy Volume](https://scholar.google.com/scholar?oi=bibs&cluster=2895622316495592648&btnI=1&hl=en)
Y Kreinin, P Gunn, D Chklovskii, J Wu - bioRxiv, 2023
- I got errors in a unit test.

```
(pystack3d) ➜  pystack3d git:(main) ✗ pytest .
===================================================================================== test session starts =====================================================================================
platform darwin -- Python 3.11.9, pytest-8.3.2, pluggy-1.5.0
rootdir: /Users/jingpengw/code/pystack3d
configfile: pyproject.toml
collected 0 items / 1 error

=========================================================================================== ERRORS ============================================================================================
_______________________________________________________________________ ERROR collecting tests/test_synthetic_stack.py ________________________________________________________________________
ImportError while importing test module '/Users/jingpengw/code/pystack3d/tests/test_synthetic_stack.py'.
Traceback:
../../anaconda3/envs/pystack3d/lib/python3.11/importlib/__init__.py:126: in import_module
tests/test_synthetic_stack.py:8: in <module>
    from examples.ex_synthetic_stack import ex_synthetic_stack as ex_synth
examples/ex_synthetic_stack.py:18: in <module>
    from utils import init_dir, postpro, plot_results, plot_cube_faces
E   ModuleNotFoundError: No module named 'utils'
=================================================================================== short test summary info ===================================================================================
ERROR tests/test_synthetic_stack.py
!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! Interrupted: 1 error during collection !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
====================================================================================== 1 error in 0.76s =======================================================================================
```
