### 2024-08-08 27ec8a6
manage NPROC and change tests references
