### 2024-08-08 #7079 @patquem
Thank you very much, @xiuliren, for your initial feedback.

> The usage requires that the image sequence be named as a specific pattern.

Given that Pystack3d's strategy is to process a stack slice by slice (saving each in separate .tif files rather than a single large '3D' tif for various reasons, such as RAM limitations, ease of multiprocessing, and visual inspection), the .tif filenames must at least include the `slice number`. The `dz` value is only needed for resampling. The parser used to extract these two pieces of information (slice number and dz) is **customizable**.
_(This format was inspired by the one used by our FIB-SEM Zeiss equipment, but feel free to contribute to the Pystack3D project by proposing your own format)._

> normal image stack, such as tif

Could you clarify what you mean by this ?

> chunks could be stitched together with significant boundary effect?

If I understand correctly, you are concerned that chunking could lead to boundary effects. This is not the case, either because the slices processesings are independent each other  or because the process requires knowledge of only 2 neighbor slices (pairing), as in the case of registration or resampling. This is handled in the processing by using an overlay of one slice in such cases. The 3D background removal, on the other hand, considers the entire downsampled stack during the pre-estimation of coefficients before applying them slice by slice.

> non-uniform textures...

Excellent point. The intensity rescaling relies on average histograms (resulting of the slice by slice histograms concatenation). In the case of non-uniformity, the histogram smoothing window must be adapted to the scale of the texture variations... up to a certain limit (fortunately, in practice, most texture changes occur over more than 10-20 slices).

> I got errors in a unit test.

This has been fixed with an additional `conftest.py` file.

I hope my explanations have clarified things for you.
Thanks again for your feedback, and I will take some time in the next few days to read your paper.
