### 2024-08-16 #7079 @xiuliren
> normal image stack, such as tif
I mean 3D image stack saved as TIF format.

for the boundary effect, if you align them only considering neighboring slices. How do you handle the accumulated drift?

for the unit test, I tried to pull the code and did not get any update.
