### 2024-08-19 #7079 @kasasxav
Hi,

Sorry for the delay, I just came back from vacation. I'll start reviewing the paper and the code.

I just quickly went through the initial checks and I have two comments already, I thought I would mention them so the authors can work on it:

- Part of the GPL license is to add a license notice in every file (https://www.gnu.org/licenses/gpl-howto.html). Could you add that? I find this very useful in case someone reuses your code.
- I would suggest the authors to move the Developer guidelines to a CONTRIBUTING.md file in the main repo. I would also suggest having a CODE_OF_CONDUCT.md. For the issue tracker, it would be nice to have default issues for example for bugs, so you can have a template that the users follow in order to explain the problem and for the developers to be able to reproduce it. This would also help categorize issues into bugs, features, etc.
- You could have automatic testing as Github actions, have you considered it?

Keep in touch,
Xavier.
