### 2024-08-19 #7079 @patquem
@xiuliren

> How do you handle the accumulated drift?

After calculating the slice-to-slice transformation matrices (`tmats`), the cumulative drift (`tmats_cumul`) is calculated just by 'summing' these matrices : `tmats_cumul = cumdot(tmats)`

> the code did not get any update

fixed (sorry, I had forgotten to push the commits)

I just took a look at your article. The referenced code, `hifiem`, indeed has many similarities with what is offered in `pystack3d`. In this sense, referencing this paper in the article is entirely justified. I note that in terms of memory management, `hifiem` was designed from the start to handle large 3D .tif files (excellent job !).
_( The first versions of `pystack3d` addressed both types of approaches (a single 3D .tif file or multiple .tif files corresponding to slices). but finally I focused on the second option, more in line with the raw data files generated by our FIB-SEM equipments.)_
