### 2024-08-19 33568f8
add CODE_OF_CONDUCT.md
