### 2024-08-19 fb17386
add CONTRIBUTING.md
