### 2024-08-19 f15849e
add 'examples' in PYTHON_PATH
