### 2024-08-19 #7079 @patquem
Thank you @kasasxav for your initial feedback.

Regarding your proposal to add a license notice in every file, I would prefer not (like numPy or scikit-image do).
Following your suggestions:
- I have added a CODE_OF_CONDUCT.md and a CONTRIBUTING.md that, notably, notifies what informations are expected for in the Issue tracker
- The automatic testing has been activated as a Github action.

Patrick
