### 2024-08-19 #7079 @kasasxav
Hi @patquem,

Alright! good for me.

I read the paper and I have some comments on it:

-  In the summary, "acquisition technic like focused ion beam scanning", shouldn't it be acquisition techniques?
- Could you cite some paper in FIB-SEM, maybe a review or somewhere which makes it clear the power of 3d reconstruction? and the fact that there is research effort invested in that direction.
- I see that you only have "Statement of need", but in that section you also mention other software. I would separate it in "Statement of need" and "State of the field" or something similar.
- Is there other python packages for 3D reconstruction that you can use? Are there other fields in which 3D reconstruction can be used? You could add this in the State of the field section.
- Also, in the statement of need you explain the components of the software, could you have that in a separated section too? Such as Implementation, Architecture, Solution description, etc.

I think that it would be nice to work on the paper by first separating it into the sections I mentioned, and then extending them a bit so that there is more of a line to the paper and the reader can be more situated on the field and convinced that this is a great solution to use! My impression is that now it goes very straight to the point.

Something I would like to see is a comparison between PyStack3D and the macros in Fiji, since you mentioned that the later doesn't use multiprocessing. It would help support your claim that this library is faster because it uses more parallelism. You could make a small script in ImageJ where you capture the time before and after the processing, and compare it by adding it into a table like the one in your appendix. You could also do that with one python package alternative to PyStack3D.

I still have to try the software and look into the technical aspects, I will let you know more comments by then.

Best regards,
Xavier.
