### 2024-08-19 #7079 @xiuliren
@patquem
1. the unit test works now. There is a bug though. The `from utils` should be `from .utils` in the `examples/ex_synthetic_stack.py`.
2. Could you update the references, including FIB-SEM paper, my paper,
Image-based correction of continuous and discontinuous non-planar axial distortion in serial section microscopy
https://www.nature.com/articles/s42003-020-0809-4
