### 2024-08-23 658d930
change rel=1e-4 to rel=1e-2 associated with test_destriping()
