### 2024-08-23 e67d3aa
add new references / split 'Statement of need' into 'Statement of field' and 'Implementation'
