+++ paper/paper.bib
@@ -1,7 +1,5 @@
 @article{Fiji,
-  author       = {Schindelin, J. and
-                  Arganda-Carreras, I. and
-                  Frise, E},
+  author       = {Schindelin, J. and Arganda-Carreras, I. and Frise, E},
   title        = {Fiji: an open-source platform for biological-image analysis},
   month        = June,
   year         = 2012,
@@ -13,9 +11,7 @@
 }
 
 @article{PyStackReg,
-  author       = {Thevenaz, P. and
-                  Ruttiman, U.E. and
-                  Unser, M.},
+  author       = {Thevenaz, P. and Ruttiman, U.E. and Unser, M.},
   title        = {A Pyramid Approach to Subpixel Registration Based on Intensity},
   month        = January,
   year         = 1998,
@@ -27,8 +23,7 @@
 }
 
 @software{pyVSNR,
-  author       = {Pavy, K. and
-                  Quéméré, P.},
+  author       = {Pavy, K. and Quéméré, P.},
   title        = {Pyvsnr 2.0.0},
   month        = february,
   year         = 2024,
@@ -39,9 +34,7 @@
 }
 
 @article{VSNR,
-  author       = {Fehrenbach, J. and
-                  Weiss , W. and
-                  Lorenzo, C.},
+  author       = {Fehrenbach, J. and Weiss , W. and Lorenzo, C.},
   title        = {Variational algorithms to remove stationary noise. Application to microscopy imaging.},
   month        = October,
   year         = 2012,
@@ -50,4 +43,61 @@
   number       = 10,
   pages        = "4420-4430",
   doi          = {10.1109/TIP.2012.2206037}
+}
+
+@article{Hoflich,
+  title={Roadmap for focused ion beam technologies},
+  author={H{\"o}flich, Katja and Hobler, Gerhard and Allen, Frances I and Wirtz, Tom and Rius, Gemma and McElwee-White, Lisa and Krasheninnikov, Arkady V and Schmidt, Matthias and Utke, Ivo and Klingner, Nico and others},
+  month        = December,
+  year         = 2023,
+  journal      = {Applied Physics Reviews},
+  volume       = 10,
+  number       = 4,
+  publisher    = {AIP Publishing},
+  doi          = {10.1063/5.0162597}
+}
+
+@article{Osenberg,
+  title        = {Classification of FIB/SEM-tomography images for highly porous multiphase materials using random forest classifiers},
+  journal      = {Journal of Power Sources},
+  volume       = {570},
+  pages        = {233030},
+  year         = {2023},
+  issn         = {0378-7753},
+  doi          = {https://doi.org/10.1016/j.jpowsour.2023.233030},
+  url          = {https://www.sciencedirect.com/science/article/pii/S0378775323004056},
+  author       = {Markus Osenberg and André Hilger and Matthias Neumann and Amalia Wagner and Nicole Bohn and Joachim R. Binder and Volker Schmidt and John Banhart and Ingo Manke},
+}
+
+@article {Kreinin,
+  author       = {Kreinin, Yuri and Gunn, Pat and Chklovskii, Dmitri and Wu, Jingpeng},
+  title        = {High-fidelity Image Restoration of Large 3D Electron Microscopy Volume},
+  year         = {2023},
+  doi          = {10.1101/2023.09.14.557785},
+  publisher    = {Cold Spring Harbor Laboratory},
[…]
