+++ paper/paper.bib
@@ -101,3 +101,13 @@
   doi          = {10.5281/zenodo.13309520},
   url          = {https://doi.org/10.5281/zenodo.13309520}
 }
+
+@INPROCEEDINGS{SIFT,
+  author       = {Lowe, D.G.},
+  booktitle    = {Proceedings of the Seventh IEEE International Conference on Computer Vision},
+  title        = {Object recognition from local scale-invariant features},
+  year         = {1999},
+  volume       = {2},
+  pages        = {1150-1157 vol.2},
+  doi          = {10.1109/ICCV.1999.790410}
+ }
