### 2024-08-28 #7079 @patquem
Hello,

The paper has been revised to incorporate the @kasasxav 's proposals.

Regarding the comparison with Fiji processing, to be honest, since the first versions of PyStack3D in our organization (~2021), we have not really conducted such processing anymore. Thomas, my colleague and co-author, has tried to relaunch some processing tasks with Fiji, and he will provide you with some information on this matter.

Regarding the citations of @xiuliren 's papers, the reference to the Hifiem paper has been included in the article. Concerning the second paper, which presents a non-agnostic registration method, it seems, in my opinion, too specific to be cited in the present paper.

Finally, regarding the practical aspects of running the tests and examples, I don't encounter any issues. However, changing `from utils` to `from .utils`, as suggested by @xiuliren , generates in my case the following error (in PyCharm) :
`ImportError: attempted relative import with no known parent package`
