### 2024-08-28 #7079 @kasasxav
Thank you @patquem , @TomDavFr ,

I think the paper looks much better now! I have two more comments for now:

- I think the Statement of need looks good, especially because you mention different fields and why 3D reconstruction is important. I see you don't have many citations though, and it would be nice to contrast the claims. I don't think it's essential though.

- Thank you for the reply @TomDavFr. In the paper you state "Unfortunately, most of these macros do not support multiprocessing, resulting in processing times that can span hours for stacks composed of several thousand images." - I feel this claim is not really justified.  Could you maybe add some of what you mentioned in the comment about automation? You could also mention that the image analysis community is widely using Python. If you want you could include those numbers too in the table, even if they are not all the methods.

Anyways I will move on to install it and test it. I have one more question though: have you considered implementing your software as a napari plugin? https://napari.org/stable/. There are tools that make it quite easy to develop the app and the GUI (https://github.com/napari/cookiecutter-napari-plugin, https://napari.org/0.4.15/guides/magicgui.html). I understand it would be more work than initially, and I don't think it's vital for this review. But I do think it would increase your visibility and potential users.
