### 2024-08-29 #7079 @xiuliren
following the comments from @TomDavFr , There is a Julia tool SeamLess that was designed to solve the similar problem of Fiji plugin. Maybe it is worth citing and mentioning as well.
https://github.com/seung-lab/SEAMLeSS
