### 2024-09-02 9b1baf2
revisit Statement of field / add Fiji processing times
