### 2024-09-02 #7079 @patquem
Hello,

I have just made some changes to the article to incorporate the @kasasxav 's comments, in particular on the Statement of Field section.
The results of the computation times obtained with Fiji have been added too in the Appendix (and supplemented with the destriping time processing). It seems to me that these data clearly demonstrate that several hours of processing were required before (doesn't it ?).
Regarding the comment about Napari, when the first PyStack versions began working in our laboratory (2021), Napari was not at its current state. With its recent advancements, it could indeed be interesting to transfer some of PyStack’s features into Napari plugins.

Regarding the mention of @xiuliren at SEAMLeSS, the link to the GitHub site does not include a README, documentation, or paper that can explain what it is about. Despite the "under active development" mention in the site, the project has not evolved for 5 years, and the lab responsible for it does not mention this project on its website. Under these conditions, it seems difficult to cite such a project.

@kasasxav : I hope you will be able to test Pystack soon. Please let me know if you encounter any issues when running the examples and tests.
