### 2024-09-02 #7079 @mstimberg
@xiuliren and @kasasxav, many thanks for your feedback so far, and thanks to @patquem for already having incorporated many of the reviewer suggestions.

@sklumpe: how are things going for you, did you have a chance to look at the software and paper? Please be reminded that the first step is to create your checklist via `@editorialbot generate my checklist`
