### 2024-09-02 CEA-MetroCarac/pystack3d#3 @kasasxav
Poetry install fails
Hi,

I am trying to install pystack3d to test it for the [review at JOSS](https://github.com/openjournals/joss-reviews/issues/7079#issuecomment-2324619118), but when I try "poetry install" it fails with:

~/github/pystack3d$ poetry install

[tool.poetry] section not found in [...]/pystack3d/pyproject.toml

I'm using linux, do you know which could be the problem?
