### 2024-09-02 #7079 @xiuliren
@patquem
The upgraded method from SEAMLess is this paper:
https://www.nature.com/articles/s41467-023-44354-0

This paper solved the alignment issue of petabyte-scale SSEM datasets.
