### 2024-09-02 #7079 @sklumpe
> @sklumpe: how are things going for you, did you have a chance to look at the software and paper? Please be reminded that the first step is to create your checklist via `@editorialbot generate my checklist`

Working on it already, thank you for the editorialbot reminder, I had forgotten about that!
