### 2024-09-03 CEA-MetroCarac/pystack3d#3 @patquem
Hi @kasasxav ,

I have (hopefully) fixed this.

Instead of :
```
pip install poetry
poetry install
```
Just do:
`pip install . `

It should work.
