### 2024-09-03 CEA-MetroCarac/pystack3d#4 @kasasxav
ModuleNotFound error
Hi,

When I try to run the tests (either from the main folder or from the tests folder) for the [review](https://github.com/openjournals/joss-reviews/issues/7079#issuecomment-2325281666), I get this error:
from examples.ex_synthetic_stack import ex_synthetic_stack as ex_synth
examples/ex_synthetic_stack.py:18: in <module>
    from utils import init_dir, postpro, plot_results, plot_cube_faces
E   ModuleNotFoundError: No module named 'utils'

I think it might be solved by adding init files.
