### 2024-09-03 CEA-MetroCarac/pystack3d#4 @kasasxav
Hi,

I see that another reviewer commented on this error and you didn't manage to fix it yet.
I think it's quite simple, I added "import utils" or "from utils import *" in the init file and it worked.

But I saw the init file is super long though, normally you only import the necessary packages. Is there a way that the code is moved to another file? And initialized when the user instanciates Stack3d?
