### 2024-09-03 CEA-MetroCarac/pystack3d#4 @kasasxav
Hi,

I see that another reviewer commented and you didn't manage to fix it. I tried and fixed it by changing:
from utils import...
to:
from examples.utils import ...

By the way the init file in pystack3d is quite large, and init files are usually just used for the imports. Is there a way that all of the code can be moved to the point when the user initialized Stack3D() ?
