### 2024-09-03 34c4c21
move code from _init_.py into stack3d.py
