+++ pystack3d/stack3d.py
@@ -1,0 +1,449 @@
+"""
+Class 'Stack3d' used to process stacks from FIB-SEM images.
+The class has 'pathdir' attribute wich is the path to the directory where the
+data is stored and 'params' attribute which is a dictionary of processing
+parameters
+"""
+import os
+import shutil
+import sys
+import time
+from importlib import import_module
+from pathlib import Path
+from threading import Thread
+from multiprocessing import Pool, Queue
+
+import numpy as np
+import matplotlib.pyplot as plt
+from tifffile import TiffFile, TiffWriter
+from tomli import load
+from tomlkit import dump
+
+from pystack3d.utils import imread_3d_skipping, get_tags
+from pystack3d.utils_multiprocessing import (worker_init, step_wrapper,
+                                             initialize_args)
+
+PROCESS_STEPS = ['cropping', 'bkg_removal',
+                 'intensity_rescaling', 'intensity_rescaling_area',
+                 'registration_calculation', 'registration_transformation',
+                 'destriping', 'resampling', 'cropping_final']
+CMAP = plt.get_cmap("tab10")
+
+plt.rcParams['savefig.dpi'] = 300
+
+
+class Stack3d:
+    """
+    Class dedicated to Stack3d workflow management
+
+    Attributes
+    ----------
+    pathdir: Path
+        Path related to the dirname where the .tif images are stored with the
+        related metadata
+    last_step_dir: str
+        Dirname related to the last step of the workflow execution.
+        For the first step, 'last_step_dir' corresponds to 'pathdir'
+
+    Arguments
+    ---------
+    pathdir: Path, optional
+        Path related to the dirname where the .tif images are stored with the
+        related metadata.
+        If None, consider the directory where the python script has been
+        executed
+    """
+
+    def __init__(self, input_name=None):
+
+        if str(input_name).endswith('.toml'):
+            self.fname_toml = input_name
+            with open(self.fname_toml, 'rb') as fid:
+                self.params = load(fid)
+                self.pathdir = self.pathdir["dirname"]
+
+        elif os.path.isdir(input_name) or input_name is None:
+            input_name = input_name or '.'
+            self.pathdir = Path(input_name)
+            fnames = list(self.pathdir.glob("*.toml"))
+            if len(fnames) == 0:
+                raise IOError(f"there is no '.toml' file in {input_name}")
+            if len(fnames) > 1:
+                raise IOError(f"there is too much '.toml' file in {input_name}")
+            self.fname_toml = fnames[0]
+
+            with open(self.fname_toml, 'rb') as fid:
+                self.params = load(fid)
+                fname = self.fname_toml.name
+                print("\n***************************************************")
+                print(f"WARNING: 'dirname' from {fname} is NOT USED")
+                print("***************************************************\n")
+        else:
+            raise IOError(f"'input_name' {input_name} is not a valid input")
+
+        if self.params["channels"] in [None, '']:
+            self.params["channels"] = ['.']
+        self.last_step_dir = self.pathdir
+
+    def __str__(self):
+        msg = f"Pathdir = {self.pathdir}\n"
+        msg += f"Channels = {self.params['channels']}"
+        return msg
+
+    def channels(self, process_step):
+        """ Return the list of channels according to the 'process_step' """
+        channels = self.params["channels"]
+        if process_step == 'registration_calculation':
+            channels = [channels[0]]
+        return channels
+
+    def process_dirname(self, process_step, channel):
+        """ Return the process dirname wrt to 'process_step' and 'channel' """
+        if process_step == 'input':
+            process_dirname = self.pathdir / channel
[…]
