### 2024-09-03 CEA-MetroCarac/pystack3d#4 @patquem
@kasasxav , I've made the changes you mentioned. Everything is fine (the tests passed). Many thanks.
