### 2024-09-03 CEA-MetroCarac/pystack3d#4 @kasasxav
ah sorry, but then if you try to run it from examples it doesn't work.

You could try:

try:
   from utils import ...
 except ModuleNotFoundError:
 from examples.utils import ...
