### 2024-09-03 CEA-MetroCarac/pystack3d#4 @patquem
My IDE is PyCharm, which automatically adds some paths to the PYTHONPATH, making it harder to spot issues like this. The try/except solution you suggested seems like a good approach. I’ll update the package accordingly
