### 2024-09-03 #7079 @kasasxav
Hi all,

I tried the software by running the tests and the example data and it all works now, I opened a couple of issues about errors I found on the way but @patquem fixed them.

I have to say I'm not an expert of FIB-SEM so I'm unaware of the technicalities of the processing algorithms, so I cannot judge as much about the scholarly effort in that sense. I do think that it will be useful for the community and it seems to be quite complete in terms of available functions and algorithms. However, I have two concerns:

- The software is very targetted into FIB-SEM data, and especially the type of data that the users use (also according to @xiuliren). For example, it would be hard to apply this software package to fluorescence microscopy data; at least I wouldn't know how to, and it feels that there are some metadata that needs to be there that might not necessarily apply in that field.
- I'm afraid that it's not very user-friendly, in the sense that users need to go in the documentation and be familiar with the API, and code their own pipelines. I really miss a GUI. That's why I thought of napari because it would be quite fast to implement a plugin and then users would be able to upload their data into napari (which already supports hdf5, tiff, etc), and then select the type of processing they want and see the results in different layers. Alternatively, it could also be done with a custom GUI. I think this is especially important in multidisciplinary fields where maybe the person that wants to process the data does not necessarily have a background in software development.

Since I'm not an expert in FIB-SEM I want to leave this to the other authors. All the other points are checked from my side.

Best,
Xavier.
