### 2024-09-03 b8ac0b9
add try/except ModuleNotFoundError
