### 2024-09-05 #7079 @sklumpe
Dear @TomDavFr & @patquem, dear all,

I have now reviewed the code and tried installation, automated tests, and application to the provided dataset and a dataset of my own. First of all, I think this is really useful piece of software that will definitely represents a great addition to the FIBSEM toolbox. The community will thank you for it! Below, I will provide a number of issues I have faced until now when working with the package and will some suggestions for improvement.

During installation, I faced issues with pyfftw on a MacBook Pro M3 chip that I still couldn’t figure out how to solve. I am able to install and run pyfftw on its own just fine. Either way, I have now installed the package on a Linux machine, which worked just fine. I am providing a couple of lines from the error message for the Mac installation below for completion, I am certain this is not an issue of pystack3d, though:

      DEBUG:__main__:Checking with includes ['fftw3.h']...no
      error: Could not find the FFTW header 'fftw3.h'
      [end of output]

  ERROR: Failed building wheel for pyfftw
Successfully built pystack3d
Failed to build pyfftw
ERROR: ERROR: Failed to build installable wheels for some pyproject.toml based projects (pyfftw)

I ran into the same issue installing poetry, but this has been solved [here] (https://github.com/CEA-MetroCarac/pystack3d/issues/3).

When applying the package to my own dataset, I faced problems with 16bit images. When converting the images to 8 bit, the software runs just fine, also on multiple CPUs. I so far couldn’t get 16 bit images to run and see the problem with a number of different datasets. The provided dataset in the assets subdirectory of the repository is, however, 16 bit. So it might be my images and I couldn’t find the issue yet. I am providing a couple of test images to see whether this is reproducible:

https://datashare.biochem.mpg.de/s/LwpwWqQOteKVSqv

On a sidenote, these are cryo-FIBSEM images, thus the weak contrast.

The error message I am getting when killing the job is (with filepaths removed):

line 41, in cropping
    outputs_saving(output_dirname, fname, img, img_res, stats)
 line 26, in outputs_saving
    save_tif(img_res2, fname, output_dirname / fname.name)
 line 45, in save_tif
    fid.write(arr, extratags=extra_tags, compression=tags[259].value)
 line 11923, in __getitem__
    raise KeyError(key)
KeyError: 259

I think the software is reasonably user-friendly, though I agree that some kind of GUI that guides through parameter setting or more detail for parameter choices in the documentation would be helpful. Moving on from the usability to the functionality, I would have a suggestion regarding the destriping algorithm. Comparing the destriping with my [own implementation](https://github.com/sklumpe/SerialFIB/blob/main/analysis/processSEM.py) based on [Spehner et al](https://doi.org/10.1016/j.jsb.2020.107528) using wavelet decomposition with the results from pystack3d / the pyVSNR, the results from wavelet decomposition visually look better to me. The idea here is to put a gaussian blur on the vertical components of the wavelet decomposition using e.g. pywavelets. It might be worthwhile to compare these two algorithms and/or provide both within the pipeline (it is really just a couple of lines of code).

Finally, regarding the manuscript, I think the state of the field would benefit from a little more detail. For example, regarding citations, I think citing Spehner et al (https://doi.org/10.1016/j.jsb.2020.107528) would be appropriate. They have described algorithms for destriping, histogram equalization and contrast enhancement for (cryo)fibsem post processing. A open software implementation of a similar approach was published [in eLife] (https://elifesciences.org/articles/70506) and is available [here] (https://github.com/sklumpe/SerialFIB/blob/main/analysis/processSEM.py). Another software package that I believe should be mentioned for the state of the art is polishEM (https://sites.google.com/site/3demimageprocessing/polishem). I assume other packages might also be out there.

Let me know if anything is unclear.

Cheers,
Sven
