### 2024-09-05 #7079 @kasasxav
Thank you for the clarification @TomDavFr, and also @sklumpe for adding a more technical perspective on the review.
As I mentioned, I don't come from a FIB-SEM background so unfortunately I couldn't evaluate as much the technical content of the algorithms but rather the general usability of the package and I hope my comments were useful to polish the paper and code.

Anyways, all good from my side then! I have checked all my boxes.

I found this other package for FIB-SEM from Janelia (https://github.com/janelia-cellmap/fibsem-tools) and it was similar in terms of the interfacing with the user. Now that @sklumpe mentioned that this software would be useful for the FIBSEM toolbox, I'm thinking that maybe this would be something good for the community, to have a package for processing FIBSEM data where the users could use algorithms from different labs, because it seems that now there are various implementations but there is not a standarized way to do the operations (such as loading data, metadata, etc). But definitely this is outside the scope of this contribution :D

Thank you for your contribution @TomDavFr & @patquem and thank you @mstimberg for the opportunity to review this paper.
Good luck with the rest of the review!

Best,
Xavier.
