### 2024-09-06 d1edcc7
bugfix: replace self.pathdir["dirname"] into Path(self.params["dirname"])
