+++ assets/toml/params_real_stack.toml
@@ -1,4 +1,4 @@
-#input_dirname = ""
+#dirname = "C:\\Users\\..."
 ind_min = 1
 ind_max = 8
 channels = ["SE2"]
