### 2024-09-06 #7079 @patquem
@xiuliren : A big thank you  for the time you dedicated to reviewing and for your feedback. Regarding the `fibsem-tools` library mentioned in your last email, I believe it's just a DataArray reloader without any actual processing algorithm associated.

@sklumpe : Thanks for your feedback and opinion in favor of pystack3d .
- Regarding the issues you encountered with fftw3 / MacBook, I came across this: https://discussions.apple.com/thread/254931084?sortBy=rank. I don't have any further insights to offer on the matter.
- I've resolved the issue with tags[259] that you encountered _(I thought this tag was always present in the .tif files, but apparently not)_.
- Regarding destriping, the wavelet decomposition-based implementation you propose is indeed a simple, fast, and effective approach in many cases. In `pystack3d` we opted for `pyvsnr` to enable oblique curtaining effects removal that don't necessarily traverse the images completely. However, I’ll keep your solution in mind for future implementations, as it would satisfy the majority of applications (simply and quickly).
- OK for citing Spehner et al.
- Regarding the reference to `SerialFIB`, I'm not sure I fully understand the connection with pystack3D. It seems to me that this application is more about controlling a FIB-SEM rather than processing stacks, Am I right?
- Lastly, concerning `polishEM`, while the paper is easy to access, the code is much less so. Could it be that its creator, initially at the `CNB` (the repository as defined in the article), left with the source code ?

Don't hesitate to follow up on any of this.
