### 2024-09-06 #7079 @sklumpe
Hi @patquem,

To your comments / questions:

1. Thanks, yes I found this too. Anyways, this is not an issue with pystack3d, just left it here as I encountered it.
2. Perfect.
3. I think it would be very useful as I had visually better results for the images provided in the previous comment.
4. Ok.
5. No need to cite SerialFIB in this at all, that was not what I was aiming at. I was just referring for the code as Spehner et al. doesn't come with an implementation and we had included code and example data in the paper (Figure 4—figure supplement 1), while obviously giving full credit to Spehner et al.
6. Yes, polishEM is published with binaries available [here](https://sites.google.com/site/3demimageprocessing/polishem) but no code. I am not sure what the policy on this would be regarding citation in JOSS as no source code is available (is that prerequisite for JOSS?) but it is still state of the field to me, so I would cite it.

Hope that clarifies things.

Best,
Sven
