### 2024-09-13 064797b
manage no '.toml' file / revisit the 'Stack3d' class docstring
