### 2024-09-13 9910542
put constraints python < 3.12 and numpy < 2
