### 2024-09-13 9a8c776
add python 3.8 and 3.9 in classifiers
