+++ pyproject.toml
@@ -17,6 +17,8 @@ keywords = ["PyStack3D", "stack", "images", "FIB-SEM", "processing",
 license = { text = "GPL v3" }
 classifiers = [
     'Programming Language :: Python :: 3',
+    'Programming Language :: Python :: 3.8',
+    'Programming Language :: Python :: 3.9',
     'Programming Language :: Python :: 3.10',
     'Programming Language :: Python :: 3.11',
     'License :: OSI Approved :: GNU General Public License v3 (GPLv3)',
