### 2024-09-18 dc92337
add destriping_from_wavelets()
