+++ pystack3d/destriping.py
@@ -4,6 +4,7 @@ Functions related to the destriping processing
 import numpy as np
 from tifffile import TiffFile
 from pyvsnr import vsnr2d
+import pywt
 
 from pystack3d.utils import outputs_saving
 from pystack3d.utils_multiprocessing import (collect_shared_array_parts,
@@ -12,9 +13,11 @@ from pystack3d.utils_multiprocessing import (collect_shared_array_parts,
 
 def destriping(fnames=None, inds_partition=None, queue_incr=None,
                maxit=100, cvg_threshold=0, filters=None,
+               wavelet_decomposition=None,
                output_dirname=None):
     """
-    Function dedicated to destriping from the VSNR algorithm.
+    Function dedicated to destriping from the VSNR algorithm (or wavelet
+    decomposition).
     For more details, see: https://github.com/CEA-MetroCarac/pyvsnr
 
     Parameters
@@ -26,34 +29,46 @@ def destriping(fnames=None, inds_partition=None, queue_incr=None,
         working in multiprocessing
     queue_incr: multiprocessing.Queue, optional
         Queue passed to the function to interact with the progress bar
+    wavelet_decomposition: dict, optional
+        Parameters provided for destriping using wavelet decomposition.
+        See `here <https://cea-metrocarac.github.io/pystack3d/doc/api/pystack3d.destriping.html#pystack3d.destriping.destriping_from_wavelets>`_
+        Example:
+            wavelet_decomposition = {'wavelet':'coif3', 'level':4, 'sigma':4}
     maxit: int, optional
         Number of maximum iterations used by the VSNR algorithm
     cvg_threshold: float, optional
         Convergence criteria to stop the VSNR iterative process, related to the
         maximum residual variation between 2 iterations
-    filters: list of dict
-        Filters defined as a list of dictionaries that specifies for each filter
-         the corresponding 'name', 'noise_level', 'sigma' and 'theta'.
+    filters: list of dict, optional
+        Filters defined as a list of dictionaries that specifies for each filter the corresponding 'name', 'noise_level', 'sigma' and 'theta'.
         See the pyvsnr documentation for more details.
         Example:
-            filters = [{'name': "Gabor", 'noise_level': 10, 'sigma': (1, 20),
-                        'theta': -10},
-                       {'name': "Gabor", 'noise_level': 40, 'sigma': (3, 40),
-                       'theta': -10}]
+            filters = [{'name': "Gabor", 'noise_level': 10, 'sigma': (1, 20), 'theta': -10}, {'name': "Gabor", 'noise_level': 40, 'sigma': (3, 40), 'theta': -10}]
     output_dirname: str, optional
         Directory pathname for process results saving
     """
     pid_0 = inds_partition[0] == 0  # first thread
+
+    wdec = wavelet_decomposition
+    wavelet = wdec['wavelet'] if 'wavelet' in wdec else None
+    level = wdec['level'] if 'level' in wdec else None
+    sigma = wdec['sigma'] if 'sigma' in wdec else None
 
     stats = []
     for fname in fnames:
         with TiffFile(fname) as tiff:
             img = tiff.asarray()
 
-        img_res = vsnr2d(img, filters,
-                         maxit=maxit,
-                         cvg_threshold=cvg_threshold,
-                         norm=False)
+        if wavelet_decomposition is not None:
+            img_res = destriping_from_wavelets(img,
+                                               wavelet=wavelet,
+                                               level=level,
[…]
