### 2024-09-18 #7079 @sklumpe
Hi @patquem ,

thanks for the suggestion.

To 1: The dependency change did indeed resolve the problem for me, so everything is working fine on that end now as well.

To 2:  I asked colleagues to compare the images and they agreed that the VSNR looks better, so it seems to just be my visual perception / preference. So all fine with me.

To 3: Thanks for pointing the original paper out to me, I was not aware of it! I agree that you should cite this one instead.

Best,,
Sven
