### 2024-09-18 505f2df
add Munch, Spehner and PolishEM citations
