+++ paper/paper.bib
@@ -111,3 +111,47 @@
   pages        = {1150-1157 vol.2},
   doi          = {10.1109/ICCV.1999.790410}
  }
+
+ @article{Munch,
+author         = {Beat M\"{u}nch and Pavel Trtik and Federica Marone and Marco Stampanoni},
+journal        = {Opt. Express},
+keywords       = {Wavelets; CCD cameras; Fourier transforms; Imaging systems; Imaging techniques; Wavelet transforms; X-ray microscopy},
+number         = {10},
+pages          = {8567--8591},
+publisher      = {Optica Publishing Group},
+title          = {Stripe and ring artifact removal with combined wavelet --- Fourier filtering},
+volume         = {17},
+month          = {May},
+year           = {2009},
+url            = {https://opg.optica.org/oe/abstract.cfm?URI=oe-17-10-8567},
+doi            = {10.1364/OE.17.008567},
+}
+
+@article{SPEHNER,
+title           = {Cryo-FIB-SEM as a promising tool for localizing proteins in 3D},
+journal         = {Journal of Structural Biology},
+volume          = {211},
+number          = {1},
+pages           = {107528},
+year            = {2020},
+issn            = {1047-8477},
+doi             = {https://doi.org/10.1016/j.jsb.2020.107528},
+url             = {https://www.sciencedirect.com/science/article/pii/S1047847720301015},
+author          = {Daniele Spehner and Anna M. Steyer and Luca Bertinetti and Igor Orlov and Lucas Benoit and Karin Pernet-Gallay and Andreas Schertel and Patrick Schultz},
+keywords        = {3-D Cryo-FIB-SEM, Immunolabeling, Image processing and machine learning, 3D reconstruction and modelling},
+}
+
+@article{polishEM,
+author          = {Fernandez, Jose-Jesus and Torres, Teobaldo E and Martin-Solana, Eva and Goya, Gerardo F and Fernandez-Fernandez, Maria-Rosario},
+title           = "{PolishEM: image enhancement in FIB–SEM}",
+journal         = {Bioinformatics},
+volume          = {36},
+number          = {12},
+pages           = {3947-3948},
+year            = {2020},
+month           = {03},
+issn            = {1367-4803},
+doi             = {10.1093/bioinformatics/btaa218},
+url             = {https://doi.org/10.1093/bioinformatics/btaa218},
+eprint          = {https://academic.oup.com/bioinformatics/article-pdf/36/12/3947/50750545/bioinformatics\_36\_12\_3947.pdf},
+}
