### 2024-09-18 8c4a215
bugfix: if wavelet_decomposition is not None
