### 2024-09-18 #7079 @patquem
Hello everyone,

I’ve just made some (final?) adjustments to the code and the paper.
Finally, the **wavelet decomposition** method has been added for destriping, and citations to `PolishEM`, `Münch `(for the wavelet decomposition algorithm), and `Spehner `have been included in the paper. Thanks to @sklumpe for validating all the review points.
In the documentation, The [TODO list](https://cea-metrocarac.github.io/pystack3d/doc/dev_notes.html#todo-list)  has been expanded. In addition to the mention of a GUI to be developed, two items have been added.
Now, only @xiuliren remains to finalize the overall review.
Hoping that this paper can be published soon (@mstimberg)
Best,
Patrick
