### 2024-09-20 #7079 @mstimberg
Many thanks for your great reviewing work, @kasasxav, @sklumpe, and @xiuliren, and thanks for having worked closely with the reviewers to take their suggestions into account @patquem

I will now do a series of final checks to prepare the manuscript for the hand-over to an editor-in-chief. @patquem, please take care of the "author tasks" in the checklist below.
