### 2024-09-20 92348c8
change legend in Figure 1
