### 2024-09-20 3ceaea2
Thanks JOSS reviewers and editors in the Acknowledgements section
