### 2024-09-20 4cbac2d
paper: fix typos
