+++ paper/paper.md
@@ -37,7 +37,7 @@ bibliography: paper.bib
 
 Three-dimensional reconstruction from 2D image stacks is a crucial technique in various scientific domains. For instance, acquisition techniques like Focused Ion Beam Scanning Electron Microscopy (FIB-SEM) leverage this approach to visualize complex structures at the nanoscale. However, creating a "clean" 3D stack often requires image corrections to remove artifacts and inconsistencies, particularly for volume segmentation, a crucial process for 3D quantitative data analysis.
 
-Here we present ``PyStack3D`` (\autoref{fig:PyStack3D}), a Python open-source library, that aimed at performing several image ‘cleaning’ tasks in the most integrated and efficient manner possible.
+Here we present ``PyStack3D`` (\autoref{fig:PyStack3D}), a Python open-source library, that aims at performing several image ‘cleaning’ tasks in the most integrated and efficient manner possible.
 
 ![a) Synthetic stack with different types of defects. b) Corrected stack. c) Ground truth.\label{fig:PyStack3D}](../doc/_static/pystack3d.png){width=85%}
 
@@ -65,29 +65,29 @@ The ``PyStack3D`` workflow is made up of multiple processing steps, specified in
 
 The processing steps currently offered by ``PyStack3D`` are:
 
-* **cropping** to reduce the image field of view to the users ROI (Region Of Interest)
+* **cropping** to reduce the image field of view to the user's ROI (Region Of Interest)
 
 * **background removal** to reduce, from 2D or 3D polynomial approximations, large-scaled brightness and contrast variations issued for instance from shadowing or charging effects in FIB-SEM images acquisition
 
-* **intensity rescaling** to homogenize the ‘gray’ intensity distribution between successive slices and smooth out abrupt intensity jumps that can occur due for instance to variations in the beam source.
+* **intensity rescaling** to homogenize the ‘gray’ intensity distribution between successive slices and smooth out abrupt intensity jumps that can occur due to, for instance, variations in the beam source.
 
-* **registration** to correct the images misalignment due to shifting, drift, rotation, … during the images acquisition (based on the ``PyStackReg`` package [@PyStackReg])
+* **registration** to correct the image misalignment due to shifting, drift, rotation, … during the image acquisition [based on the ``PyStackReg`` package @PyStackReg]
 
-* **destriping** to minimize artefacts like stripes or curtains effects typically found in FIB-SEM images (based on the ``PyVSNR`` package [@pyVSNR], [@VSNR] or wavelet decomposition [@Munch])
+* **destriping** to minimize artefacts like stripes or curtain effects typically found in FIB-SEM images (based on the ``PyVSNR`` package [@pyVSNR], [@VSNR] or wavelet decomposition [@Munch])
 
 * **resampling** to correct non-uniform spatial inter-slice distances and enable correct 3D volume reconstructions
 
 * **final cropping** to eliminate artefacts potentially produced near the edges during the image processing or to select another ROI at the end.
 
-At the end of each process step, ``PyStack3D`` provides statistical profiles like evolution of minimum, maximum, and mean values for each slice, and relevant visualizations specific to the processing performed. In addition, 3D and 2D plots (cut-planes) akin to those shown in (\autoref{fig:PyStack3D}) and (\autoref{fig:workflow}) respectively can be produced.
+At the end of each process step, ``PyStack3D`` provides statistical profiles like evolution of minimum, maximum, and mean values for each slice, and relevant visualizations specific to the processing performed. In addition, 3D and 2D plots (cut-planes) akin to those shown in \autoref{fig:PyStack3D} and \autoref{fig:workflow}, respectively, can be produced.
 
[…]
