### 2024-09-20 10ef86e
paper: various reference/bibliography fixes
Add reference for ImageJ
