### 2024-09-20 #7079 @patquem
Hi @mstimberg

**OK** : Double check authors and affiliations (including ORCIDs)
**OK** version 2024.2 : Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.
**OK** https://doi.org/10.5281/zenodo.13808737 : Archive the release on Zenodo/figshare/etc and post the DOI here.
**OK** : Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.
**OK** :  Make sure that the license listed for the archive is the same as the software license.

(please re-edit the article for taking into account your last corrections and the zip removal)
