### 2024-09-23 #7079 @mstimberg
@patquem Apologies for delaying things again, but I had to check back with other editors for what I realized is quite ambiguous in our instructions/documentation: the archive that you upload to Zenodo should not be the "release" in the sense of what you'd upload to PyPI, but rather the **archived repository** (e.g. the zip file from here: https://github.com/CEA-MetroCarac/pystack3d/releases/tag/2024.2). The reason is that this is supposed to reflect what the reviewers looked at, and in your case the files on Zenodo do not include the examples and the documentation. If I am not mistaken, you cannot add files to an existing archive on Zenodo, so unfortunately I think you'll have to make a new release. Thanks, and apologies for the unclear instructions :pray:
