### 2024-09-23 #7079 @patquem
@mstimberg
I have done a new Zenodo version with the 'full' zip archives : [10.5281/zenodo.13827928](https://doi.org/10.5281/zenodo.13827928)
Hope it will be good 🚀
