### 2024-09-23 #7079 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
