### 2024-09-23 #7079 @Kevin-Mattheus-Moerman
@patquem as AEiC for JOSS I will now help to process this submission for acceptance in JOSS. I will now process some final checks:

- [x] Software license tag listed here matches a tagged release
