### 2024-09-24 6a647f7
add CITATION.cff and JOSS tag
