+++ CITATION.cff
@@ -1,0 +1,32 @@
+cff-version: "1.2.0"
+authors:
+- family-names: Quéméré
+  given-names: Patrick
+  orcid: "https://orcid.org/0009-0008-6936-1249"
+- family-names: David
+  given-names: Thomas
+  orcid: "https://orcid.org/0000-0002-6796-220X"
+doi: 10.5281/zenodo.13827928
+message: If you use this software, please cite our article in the
+  Journal of Open Source Software.
+preferred-citation:
+  authors:
+  - family-names: Quéméré
+    given-names: Patrick
+    orcid: "https://orcid.org/0009-0008-6936-1249"
+  - family-names: David
+    given-names: Thomas
+    orcid: "https://orcid.org/0000-0002-6796-220X"
+  date-published: 2024-09-23
+  doi: 10.21105/joss.07079
+  issn: 2475-9066
+  issue: 101
+  journal: Journal of Open Source Software
+  publisher:
+    name: Open Journals
+  start: 7079
+  title: "PyStack3D: A python package for fast image stack correction"
+  type: article
+  url: "https://joss.theoj.org/papers/10.21105/joss.07079"
+  volume: 9
+title: "PyStack3D: A python package for fast image stack correction"
